# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Commission Calculation System API
  version: "1.0.0"
  description: >
    REST API that powers the Commission Calculation System. The service exposes
    endpoints for authentication, agent management, sales capture, dashboards,
    reporting, and clawback operations.
servers:
  - url: http://localhost:5002
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Policy:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
        product:
          type: string
        fyc_rate:
          type: number
          format: float
...
    ReportSummary:
      type: object
      properties:
        period:
          type: object
          properties:
            mode:
              type: string
            label:
              type: string
            start:
              type: string
            end:
              type: string
        by_agent_bar:
          type: array
          items:
            type: object
            properties:
              agent:
                type: string
              commission:
                type: number
        trend:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              value:
                type: number
        kpi_total:
          type: number
        history:
          type: array
          items:
            type: object
            properties:
              quarter:
                type: string
              agent:
                type: string
              total_sale:
                type: number
              avg_rate:
                type: number
              total_commission:
                type: number
        volume_bonus:
          type: array
          items:
            type: object
            properties:
              agent:
                type: string
              volume:
                type: number
              tier:
                type: string
              bonus:
                type: number
        agents:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
    ClawbackSummary:
      type: object
      properties:
        pending:
          type: integer
        avg:
          type: number
        impact:
          type: number
        trend:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              value:
                type: number
        by_sales:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
    ClawbackList:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              policy_number:
                type: string
              cancellation_date:
                type: string
                format: date
              status:
                type: string
              amount:
                type: number
              original_amount:
                type: number
              salesperson:
                type: string
    ClawbackDetail:
      type: object
      properties:
        id:
          type: integer
        policy:
          $ref: '#/components/schemas/Policy'
        cancellation_date:
          type: string
          format: date
        status:
          type: string
        reason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              agent:
                type: string
              agent_id:
                type: integer
              entry_type:
                type: string
              original_amount:
                type: number
              clawback_amount:
                type: number
              meta:
                type: object
                nullable: true
    ClawbackPreview:
      type: object
      properties:
        exists:
          type: boolean
        message:
          type: string
          nullable: true
        policy:
          $ref: '#/components/schemas/Policy'
        first_sale_date:
          type: string
          format: date
        rule_factor:
          type: number
        totals:
          type: object
          properties:
            original:
              type: number
            clawback:
              type: number
        items:
          type: array
          items:
            type: object
            properties:
              agent_id:
                type: integer
              entry_type:
                type: string
              original_amount:
                type: number
              clawback_amount:
                type: number
              meta:
                type: string
                nullable: true
      required:
        - exists
paths:
...
  /reports/summary:
    get:
      tags:
        - Reports
      security:
        - bearerAuth: []
      summary: Commission report summary for a selected period
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [monthly, quarterly, annual]
            default: quarterly
        - in: query
          name: ref
          schema:
            type: string
            description: Optional reference period (YYYY-MM or YYYY)
        - in: query
          name: agent_id
          schema:
            type: string
            description: Agent ID or "all"
      responses:
        '200':
          description: Report payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSummary'
  /reports/export:
    get:
      tags:
        - Reports
      security:
        - bearerAuth: []
      summary: Export summary data as a CSV attachment
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [monthly, quarterly, annual]
        - in: query
          name: ref
          schema:
            type: string
        - in: query
          name: agent_id
          schema:
            type: string
      responses:
        '200':
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /clawbacks/summary:
    get:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: Aggregated statistics for clawback monitoring
      responses:
        '200':
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClawbackSummary'
  /clawbacks:
    get:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: List clawbacks with totals
      parameters:
        - in: query
          name: q
          schema:
            type: string
            description: Filter by policy number or salesperson name
      responses:
        '200':
          description: Clawback list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClawbackList'
    post:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: Create a clawback request from a preview payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policy_number:
                  type: string
                cancellation_date:
                  type: string
                  format: date
                reason:
                  type: string
                  nullable: true
                notes:
                  type: string
                  nullable: true
              required:
                - policy_number
                - cancellation_date
      responses:
        '200':
          description: Clawback created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  status:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clawbacks/{clawbackId}:
    get:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: Get details for a specific clawback
      parameters:
        - in: path
          name: clawbackId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Clawback detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClawbackDetail'
        '404':
          description: Clawback not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clawbacks/preview:
    get:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: Preview recoverable commissions prior to creating a clawback
      parameters:
        - in: query
          name: policy_number
          required: true
          schema:
            type: string
        - in: query
          name: cancellation_date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Preview payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClawbackPreview'
        '400':
          description: Missing parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clawbacks/approve:
    post:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: Bulk approve clawback requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
              required:
                - ids
      responses:
        '200':
          description: Approval results
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        status:
                          type: string
  /clawbacks/deny:
    post:
      tags:
        - Clawbacks
      security:
        - bearerAuth: []
      summary: Bulk deny clawback requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
              required:
                - ids
      responses:
        '200':
          description: Denial results
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        status:
                          type: string